requirements.dev.txt
PyYAML
boto3

s3pypi
wheel